第四天提交

+++ model/src/main/java/com/atguigu/gmall/product/entity/SkuSaleAttrValue.java
@@ -28,7 +28,7 @@ public class SkuSaleAttrValue implements Serializable {
     /**
      * spu_id(冗余)
      */
-    private Integer spuId;
+    private Long spuId;
 
     /**
      * 销售属性值id
